improvement: Cloning: improve seek accuracy
By doing `setTimeout` as late as possible. In theory if `setTimeout` gets
called way before the silence start, it may be not ideally in sync with
`element.currentTime`. But what I have managed to observe is only that
`currentTime - seekScheduledTo` would sometimes be roughly -0.1, which is like
decreasing marginAfter by this value.

This also simplifies the code a bit