improvement: support Firefox mobile (Android)
and iOS maybe - did not test.
Should also make things better on the Kiwi browser -
did not test either.

On YouTube it works terribly though.
It refuses to attach to the element until you reload the page
and Retry searching for the element.
Also the video pauses when you open the popup,
so the chart is useless.

This issue doesn't appear happen on other websites (e.g. Coursera)
and the local file player.