fix: `_lastEmitedEventIsSilenceStartEvent = false`
The code that uses the processor expects the first event
to always be `SILENCE_START`.
It's not correct to say that we did emit the `SILENCE_START`
event when we didn't actually emit it.

It might get incorrectly set to `true` if `durationThreshold` is 0,
which is when margins are 0.

And this `initialDuration` parameter is not even used.